pyproject.toml
[tool.poetry]
name = "django-rest-framework-protobuf"
version = "0.1.0-alpha"
description = "Protobuf support for django-rest-framework"
license = "MIT"
authors = ["Denis Cornehl <denis.cornehl@gmail.com>"]

repository = "https://github.com/syphar/django-rest-framework-poetry/"
homepage = "https://github.com/syphar/django-rest-framework-poetry/"

packages = [
    { include = "rest_framework_protobuf" },
]


[tool.poetry.dependencies]
python = "^3.7"
protobuf = "^3.11.3"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
pydocstyle = "^5.0.2"
flake8 = "^3.7.9"
isort = "^4.3.21"
bandit = "^1.6.2"
pre-commit = "^2.0.1"
invoke = "^1.4.1"
pytest-django = "^3.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
